Shuffle some code around
